clean up a bit and break up mesh strategies
