+++ ajc27_freemocap_blender_addon/core_functions/meshes/skelly_mesh/helpers/empty_markers_for_rest_pose.py
@@ -1,4 +1,5 @@
 # TODO: merge with a main empty information dictionary
+
 _EMPTY_MARKERS = {
     'hips_center': {
         "bone": 'pelvis',
